fix services and add edpoint which can change roles

+++ src/services/ProductCoffeeServiceImpl.ts
@@ -2,7 +2,8 @@ import {ProductCoffeeService} from "./ProductCoffeeService";
 import {Coffee} from "../model/Coffee";
 import {CoffeeDto} from "../model/CoffeeDto";
 import {configuration} from "../config/config";
-import {CoffeeQuantity} from "../utils/types";
+import {CoffeeQuantity, CoffeeReturn} from "../utils/types";
+import {ResultSetHeader} from "mysql2";
 
 
 
@@ -12,9 +13,6 @@ export class ProductCoffeeServiceImpl implements ProductCoffeeService {
     async changeQuantity(name:string, count:number) {
         await configuration.pool.query("UPDATE products_coffee SET quantity =quantity-? WHERE name = ?;",[count,name]);
     }
-
-
-
     async addCoffee(coffee: Coffee): Promise<void> {
 
         try{
@@ -25,27 +23,30 @@ export class ProductCoffeeServiceImpl implements ProductCoffeeService {
         }
     }
     async changeCoffee(id: string, coffee:CoffeeDto): Promise<boolean> {
-        const [result] = await configuration.pool.query(
-            "UPDATE products_coffee SET name=?, price =?, quantity=?, status=? WHERE id=?",
+        const [result] = await configuration.pool.query<ResultSetHeader>(
+            "UPDATE products_coffee SET name=?, price =?, quantity=? WHERE id=?",
             [coffee.name, coffee.price, coffee.quantity,  id]
         );
-        //todo
-        return Promise.resolve(!result ? false : true);
+
+        return Promise.resolve(result.changedRows > 0  );
     }
     async quantityCoffeeByName(name: string): Promise<CoffeeQuantity> {
         const [result] = await configuration.pool.query<CoffeeQuantity[]>("SELECT name, quantity FROM products_coffee WHERE name= ?",[name]);
         if(!result[0]) throw new Error(JSON.stringify({status: 404,message:`No product with name ${name} found.`}))
         return Promise.resolve(result[0]);
     }
-    async getAllCoffees(): Promise<Coffee[]> {
-        const [result] = await configuration.pool.query("SELECT * FROM products_coffee")
-        return Promise.resolve(result as Coffee[]);
+    async getAllCoffees(): Promise<CoffeeReturn[]> {
+        const [result] = await configuration.pool.query<CoffeeReturn[]>("SELECT * FROM products_coffee")
+        return Promise.resolve(result);
     }
     async getCoffeeByName(name: string): Promise<Coffee> {
-        const [result] = await configuration.pool.query("SELECT * FROM products_coffee WHERE name=?",[name])
-        return Promise.resolve((result as Coffee[])[0]);
+        const [result] = await configuration.pool.query<CoffeeReturn[]>("SELECT * FROM products_coffee WHERE name=?",[name])
+        return Promise.resolve(result[0]);
     }
     async removeCoffee(id: string): Promise<void> {
+        const [result] = await configuration.pool.query<CoffeeReturn[]>("SELECT * FROM products_coffee WHERE id=?",[id])
+        if(!result[0]) throw new Error(JSON.stringify({status:404,message:"Coffee not found"}))
+
         await configuration.pool.query("DELETE FROM products_coffee WHERE id=?",[id]);
     }
 }
